employee and cases operation
employee and cases operation

+++ App/Models/Common_m.php
@@ -47,16 +47,56 @@ class Common_m extends Model
         return FALSE;
         //it will be return boolean value (true/false)
     }
+    
+    
+     public function register_employee($params){        
+        $email_exist = $this->db->query("SELECT * FROM employee WHERE user_email_id='".$params['user_email_id']."' ");
+        $res = $email_exist->getRowArray();                       
+        if($res){
+            return Array(
+                'success' => false,
+                'email_exist' => true
+            );
+        }    
+        $params['user_email_password'] = md5($params['user_email_password']);        
+        unset($params['g-recaptcha-response']);     
+    
+        $params['created_by']=$_SESSION['admin']['admin_user_id'];
+        
+        $builder = $this->db->table('employee');
+        $builder->insert($params);
+        $insert_id = $this->db->insertID();
+        
+        if(!empty($insert_id)){
+            return Array(
+                'success' => true,
+                'email_exist' => false,                
+                'email' => $params['user_email_id'],
+                'employee_user_id'=>$insert_id
+            );
+        }
+        return FALSE;
+        //it will be return boolean value (true/false)
+    }
+    
     public function chek_code_exist($user_id,$link_code,$user_type) {       
         $q = "SELECT * FROM user_email_link WHERE user_id=$user_id AND user_type='$user_type' AND link_code='$link_code' ";
         $query = $this->db->query($q);
         $row = $query->getRowArray(); 
         if (isset($row))
         {
-            $table="hpshrc_customer";
-            $this->db->query("UPDATE $table SET customer_status='ACTIVE',customer_attempt=0,customer_locked_status=0,customer_email_verified_status=1 WHERE customer_id=$user_id ");
-            $this->db->query("DELETE FROM user_email_link WHERE user_id=$user_id AND user_type='$user_type' AND link_code='$link_code'");
-            return true;
+            if($user_type=='customer'){
+                $table="hpshrc_customer";
+                $this->db->query("UPDATE $table SET customer_status='ACTIVE',customer_attempt=0,customer_locked_status=0,customer_email_verified_status=1 WHERE customer_id=$user_id ");
+                $this->db->query("DELETE FROM user_email_link WHERE user_id=$user_id AND user_type='$user_type' AND link_code='$link_code'");
+                return true;
+            }
+            if($user_type=='employee'){
+                $table="employee";
+                $this->db->query("UPDATE $table SET user_status='ACTIVE',user_attempt=0,user_locked_status=0,user_email_verified_status=1 WHERE employee_user_id=$user_id ");
+                $this->db->query("DELETE FROM user_email_link WHERE user_id=$user_id AND user_type='$user_type' AND link_code='$link_code'");
+                return true;
+            }
         }
         return false;
     }
@@ -91,6 +131,33 @@ class Common_m extends Model
             );
         }        
         return FALSE;
+    }
+    
+      public function edit_employee($params,$employee_id){  
+        $email_exist = $this->db->query("SELECT * FROM employee WHERE user_email_id='".$params['user_email_id']."' AND employee_user_id!={$employee_id} ");
+        $res = $email_exist->getRowArray();                       
+        if($res){
+            return Array(
+                'success' => false,
+                'email_exist' => true
+            );
+        }          
+        $builder = $this->db->table('employee');
+        $builder->where('employee_user_id', $employee_id);
+        $update =$builder->update($params);  
+        
+        if($update){
+            return Array(
+                'success' => true,
+                'email_exist' => false               
+            );
+        }        
+        return FALSE;
+    }
+     public function get_single_employee($employee_id){  
+        $res = $this->db->query("SELECT * FROM employee WHERE employee_user_id={$employee_id} ");
+        $row = $res->getRowArray();
+        return $row;
     }
     public function get_single_customer($customer_id){  
         $res = $this->db->query("SELECT * FROM hpshrc_customer WHERE customer_id={$customer_id} ");
